+++ Lab1/HelpFunctions.h
@@ -1,0 +1,8 @@
+#ifndef HELPFUNCTIONS_H_INCLUDED
+#define HELPFUNCTIONS_H_INCLUDED
+double input (double &a, double &b, double &c);
+double discriminant(double a,double b,double c);
+void solution(double a,double b,double c);
+
+
+#endif
